Remove table row
